corp-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pymupdf==1.23.8
pdfplumber==0.10.3
Pillow==10.1.0
openai==1.3.0 